+++ auth/model/token/model.go
@@ -15,83 +15,96 @@ const (
 
 var ErrorEmptyToken = errors.New("Token string cannot be empty")
 var ErrorExpiredToken = errors.New("Token has expired")
-var ErrorEmptyUsersTable = errors.New("users table cannot be empty")
+var ErrorNilQuitChanel = errors.New("Quit channel cannot be nil")
 var ErrorBadUserID = errors.New("UserID cannot be empty")
+var ErrorGCRunning = errors.New("Garbage collector is already running")
 
 type Model struct {
-	db               *sql.DB
-	usersTable_IDCol string
+	db *sql.DB
 
-	insCh chan token
-	delCh chan string
+	gcRunning bool
+	insCh     chan token
+	delCh     chan string
 }
 
-func NewModel(db *sql.DB, usersTable_IDCol string, quitCh chan error) (*Model, error) {
+func NewModel(db *sql.DB) (*Model, error) {
 
 	if db == nil {
 		return nil, helper.ErrorNilDB
-	}
-
-	if usersTable_IDCol == "" {
-		return nil, ErrorEmptyUsersTable
 	}
 
 	iCh := make(chan token)
 	dCh := make(chan string)
 
 	m := &Model{
-		db:               db,
-		usersTable_IDCol: usersTable_IDCol,
-		insCh:            iCh,
-		delCh:            dCh,
+		db:    db,
+		insCh: iCh,
+		delCh: dCh,
 	}
 
-	go m.garbageCollect(quitCh)
-
 	return m, nil
+}
+
+func (m *Model) RunGarbageCollector(quitCh chan error) error {
+
+	if quitCh == nil {
+		return ErrorNilQuitChanel
+	}
+
+	smallest, err := m.GetSmallestExpiry()
+	if err != nil {
+		return err
+	}
+
+	go m.garbageCollect(smallest, quitCh)
+	return nil
 }
 
 func (m *Model) TableName() string {
 	return tableName
 }
 
-func (m Model) PrimaryKeyField() string {
+func (m *Model) PrimaryKeyField() string {
 	return tableName + ".id"
 }
 
 func (m *Model) TableDesc() string {
+	// TODO enforce fk integrity (userID)
+	// CHECK https://github.com/cockroachdb/cockroach/issues/2132
 	return fmt.Sprintf(`
 		id	SERIAL		PRIMARY KEY,
-		userID	INT		NOT NULL CHECK (userID IN %s),
+		userID	INT		NOT NULL,
 		devID	STRING		NOT NULL,
 		token	STRING		UNIQUE NOT NULL,
-		issue	TIMESTAMP	NOT NULL,
+		issued	TIMESTAMP	NOT NULL,
 		expiry	TIMESTAMP	NOT NULL,
 		INDEX userIDIndex (userID)
-	`, m.usersTable_IDCol)
+	`)
 }
 
 func (m *Model) Save(t token) (int, error) {
 
 	qStr := fmt.Sprintf(`
-		INSERT INTO %s (userID, devID, token, issue, expiry)
+		INSERT INTO %s (userID, devID, token, issued, expiry)
 		VALUES ($1, $2, $3, $4, $5)
 		 RETURNING id
 		`, tableName)
 
 	var tokenID int
 	err := m.db.QueryRow(qStr, t.userID, t.devID, t.token,
-		t.issue, t.expiry).Scan(&tokenID)
+		t.issued, t.expiry).Scan(&tokenID)
 	if err != nil {
 		return tokenID, err
 	}
 
 	t.id = tokenID
-	m.insCh <- t
+	if m.gcRunning {
+		m.insCh <- t
+	}
 	return tokenID, nil
 }
 
-func (um Model) Get(usrID int, tknStr string) (*token, error) {
+func (um *Model) Get(usrID int, tknStr string) (*token, error) {
 
 	if usrID < 1 {
 		return nil, ErrorBadUserID
@@ -102,7 +115,7 @@ func (um Model) Get(usrID int, tknStr string) (*token, error) {
 	}
 
 	qStr := fmt.Sprintf(`
-		SELECT id, userID, devID, token, issue, expiry
+		SELECT id, userID, devID, token, issued, expiry
 		FROM %s
 		WHERE userID = $1
 		AND token = $2
@@ -110,39 +123,40 @@ func (um Model) Get(usrID int, tknStr string) (*token, error) {
 
 	tkn := &token{}
 	err := um.db.QueryRow(qStr, usrID, tknStr).Scan(
-		&tkn.id, &tkn.userID, &tkn.devID, &tkn.token, &tkn.issue, &tkn.expiry,
+		&tkn.id, &tkn.userID, &tkn.devID, &tkn.token, &tkn.issued, &tkn.expiry,
 	)
+	if err != nil {
+		if err.Error() == helper.NoResultsErrorStr {
+			return nil, nil
+		}
+		return nil, err
+	}
+
+	tknI, err := um.ValidateExpiry(tkn)
 	if err != nil {
 		return nil, err
 	}
 
-	if time.Now().After(tkn.expiry) {
-
-		ok, err := um.Delete(tknStr)
-		if err != nil {
-			return nil, fmt.Errorf("%s ...further %s while deleting the token",
-				ErrorExpiredToken, err)
-		}
-
-		if !ok {
-			return nil, fmt.Errorf("%s ...further, token was not deleted successfully",
-				ErrorExpiredToken)
-		}
-
-		return nil, ErrorExpiredToken
+	tkn, ok := tknI.(*token)
+	if !ok {
+		return nil, fmt.Errorf("Error handling token type got %T expected %T", tkn, &token{})
 	}
 
 	return tkn, nil
 }
 
-func (um Model) Delete(token string) (bool, error) {
+func (um *Model) Delete(tknStr string) (bool, error) {
+
+	if tknStr == "" {
+		return false, ErrorEmptyToken
+	}
 
 	qStr := fmt.Sprintf(`
 		DELETE FROM %s
 		WHERE token = $1
 	`, tableName)
 
-	r, err := um.db.Exec(qStr, token)
+	r, err := um.db.Exec(qStr, tknStr)
 	if err != nil {
 		return false, err
 	}
@@ -160,25 +174,88 @@ func (um Model) Delete(token string) (bool, error) {
 		return false, nil
 	}
 
-	um.delCh <- token
+	if um.gcRunning {
+		um.delCh <- tknStr
+	}
 	return true, nil
 }
 
-func (um Model) garbageCollect(quitCh chan error) {
+func (um *Model) ValidateExpiry(tkn Token) (Token, error) {
 
-	smallest, err := um.getSmallestExpiry()
+	if time.Now().Before(tkn.Expiry()) {
+		return tkn, nil
+	}
+
+	_, err := um.Delete(tkn.Token())
 	if err != nil {
-		quitCh <- err
+		return nil, fmt.Errorf("%s ...further %s while deleting the token",
+			ErrorExpiredToken, err)
+	}
+
+	return nil, ErrorExpiredToken
+}
+
+func (um *Model) GetSmallestExpiry() (*token, error) {
+
+	qStr := fmt.Sprintf(`
+		SELECT id, userID, devID, token, issued, expiry
+		FROM %s
+		WHERE expiry = (SELECT MIN(expiry) FROM %s)
+	`, tableName, tableName)
+
+	t := &token{}
+	err := um.db.QueryRow(qStr).Scan(
+		&t.id, &t.userID, &t.devID, &t.token, &t.issued, &t.expiry,
+	)
+
+	if err != nil {
+		if err.Error() == helper.NoResultsErrorStr {
+			return nil, nil
+		}
+		return nil, err
+	}
+
+	return t, nil
+}
+
+// garbageCollect deletes all expired tokens.
+// quitCh - sends an error value on quit (nil if no error occurred).
+// garbageCollect is best run in a goroutine as it blocks in a
+// forever loop until one of the following signals is received:
+// 1. Token with earliest expiry time has just expired (garbage collect it).
+// 2. A token has been inserted (recalculate token with earliest expiry time).
+// 3. A token has been deleted (recalculate token with earliest expiry time).
+func (um *Model) garbageCollect(smallest *token, quitCh chan error) {
+
+	if quitCh == nil {
 		return
 	}
 
-	nextExp := smallest.expiry.Sub(time.Now())
+	if um.gcRunning {
+		quitCh <- ErrorGCRunning
+		return
+	}
+
+	um.gcRunning = true
+	defer func() { um.gcRunning = false }()
+
+	nextExp := longDuration
+	if smallest != nil {
+		nextExp = smallest.expiry.Sub(time.Now())
+	}
+
 	timer := time.NewTimer(nextExp)
 
+loop:
 	for {
 		select {
 
 		case <-timer.C:
+			if smallest == nil {
+				continue loop
+			}
+			// um.Delete() should send to delCh on success which will
+			// trigger calculation of the next smallest
 			_, err := um.Delete(smallest.token)
 			if err != nil {
 				quitCh <- err
@@ -186,7 +263,7 @@ func (um Model) garbageCollect(quitCh chan error) {
 			}
 
 		case inserted := <-um.insCh:
-			if smallest.expiry.After(inserted.expiry) {
+			if smallest == nil || smallest.expiry.After(inserted.expiry) {
 				smallest = &inserted
 			}
 
@@ -195,33 +272,22 @@ func (um Model) garbageCollect(quitCh chan error) {
 				continue
 			}
 
-			smallest, err = um.getSmallestExpiry()
+			var err error
+			smallest, err = um.GetSmallestExpiry()
 			if err != nil {
 				quitCh <- err
 				return
 			}
 
+			nextExp = shortDuration
+			if smallest != nil {
+				nextExp = smallest.expiry.Sub(time.Now())
+			}
+
+			timer = time.NewTimer(nextExp)
+
 		}
 	}
 
 	quitCh <- nil
-}
-
-func (um Model) getSmallestExpiry() (*token, error) {
-
-	qStr := fmt.Sprintf(`
-		SELECT id, userID, devID, token, issue, expiry
-		FROM %s
-		WHERE expiry = (SELECT MIN(date) FROM %s)
-	`, tableName, tableName)
-
-	var t *token
-	err := um.db.QueryRow(qStr).Scan(
-		&t.id, &t.userID, &t.devID, &t.token, &t.issue, &t.expiry,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return t, nil
 }
